requirements.txt
appnope==0.1.0
beautifulsoup4==4.6.0
cffi==1.1.2
cryptography==0.9.1
cssselect==1.0.3
decorator==4.0.4
enum34==1.0.4
feedparser==5.1.3
Flask==0.10.1
gevent==1.0.1
gnureadline==6.3.3
google-api-python-client==1.3.1
greenlet==0.4.5
httplib2==0.9
idna==2.0
ipaddress==1.0.7
ipython==4.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.7.3
locustio==0.7.2
lxml==4.1.1
MarkupSafe==0.23
msgpack-python==0.4.5
numpy==1.9.2
oauth2client==1.4.6
pandas==0.17.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.14
pygerduty==0.28
pymongo==3.0.2
pyOpenSSL==0.15.1
python-dateutil==2.4.2
pytz==2015.4
redis==2.10.6
requests==2.5.3
rfc3339==5
rsa==3.1.4
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
simplejson==3.6.5
six==1.9.0
traitlets==4.0.0
ua-parser==0.7.3
uritemplate==0.6
user-agents==1.1.0
Werkzeug==0.10.1
